mcp-server/backend/requirements.txt
Flask
google-cloud-aiplatform
gunicorn
google-cloud-storage
google-auth
Flask-Cors
